+++ src/main/java/fr/antoinenebout/vacation_app_api/dto/Vacation/VacationSummaryDTO.java
@@ -8,7 +8,7 @@ import java.util.Date;
 public class VacationSummaryDTO {
 
     private Long id;
-    private Long requested;
+    private Double requested;
 
     private String type;
     private String group;
